+++ changes.md
@@ -3,6 +3,7 @@ Change Log
 
 ### changes in 0.2-SNAPSHOT
 
+* moved loadBundle(source: InputStream) from MarkerSupport to Neo4jFileLoader
 
 ### changes in 0.1